+++ apps/web/components/layout/Sidebar/index.tsx
@@ -1,13 +1,12 @@
 'use client'
 
 import { useState } from 'react'
-import { SidebarOpen } from 'lucide-react'
-import { Button } from '@rdx/ui/components/button'
 import { usePathname } from 'next/navigation'
 import type { DocCategory } from '@rdx/types/loaders'
 
 import { DesktopSidebar } from './Desktop'
 import { MobileSidebar } from './Mobile'
+import { SidebarToggle } from './toggle'
 
 export const Sidebar = ({ categories }: { categories: DocCategory[] }) => {
   const pathname = usePathname()
@@ -15,17 +14,13 @@ export const Sidebar = ({ categories }: { categories: DocCategory[] }) => {
 
   return (
     <>
-      <div className="md:hidden absolute top-2 left-4 z-50">
-        <Button variant="ghost" onClick={() => setIsOpen(!isOpen)}>
-          <SidebarOpen className="size-4" />
-        </Button>
-      </div>
+      <SidebarToggle onToggle={() => setIsOpen(!isOpen)} />
 
       <aside className="h-screen w-auto md:w-40 lg:w-56 flex-shrink-0 transition-all">
         <div className="h-full hidden md:block">
           <DesktopSidebar pathname={pathname} categories={categories} />
         </div>
-        <div className="w-full md:hidden block relative hidden">
+        <div className="w-full md:hidden block relative">
           <MobileSidebar
             pathname={pathname}
             categories={categories}